backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
pydantic
pydantic-settings
python-dotenv
google-genai
requests
